+++ packages/preact/src/methods/submit.ts
@@ -3,11 +3,11 @@ import type { FieldValues, FormStore, ResponseData } from '../types';
 /**
  * Validates and submits the form.
  *
- * @param form The form to be validated and submitted.
+ * @param form The form to be submitted.
  */
 export function submit<
   TFieldValues extends FieldValues,
   TResponseData extends ResponseData
 >(form: FormStore<TFieldValues, TResponseData>): void {
-  form.element.value?.requestSubmit();
+  form.element.peek()?.requestSubmit();
 }