+++ packages/qwik/src/methods/submit.ts
@@ -3,7 +3,7 @@ import type { FieldValues, FormStore, ResponseData } from '../types';
 /**
  * Validates and submits the form.
  *
- * @param form The form to be validated and submitted.
+ * @param form The form to be submitted.
  */
 export function submit<
   TFieldValues extends FieldValues,