+++ packages/solid/src/methods/submit.ts
@@ -1,13 +1,14 @@
+import { untrack } from 'solid-js';
 import type { FieldValues, FormStore, ResponseData } from '../types';
 
 /**
  * Validates and submits the form.
  *
- * @param form The form to be validated and submitted.
+ * @param form The form to be submitted.
  */
 export function submit<
   TFieldValues extends FieldValues,
   TResponseData extends ResponseData
 >(form: FormStore<TFieldValues, TResponseData>): void {
-  form.element?.requestSubmit();
+  untrack(() => form.element)?.requestSubmit();
 }